basic modal created to create competition
