+++ frontend/src/Components/Header/Header.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import classes from "./Header.module.css";
 import Button from "../../Components/Buttons/Button";
+import CreateCompetition from "../Modals/CreateCompetition";
 
 function Header(props) {
     let headerContainer = classes.headerContainer;
@@ -12,12 +13,18 @@ function Header(props) {
             <div className={classes.textContainer}>
                 <h1>Olá {props.organization} 👋</h1>
             </div>
-            <Button
-                text={"Criar prova"}
-                onClick={() => {
-                    /* TODO: openModal */
-                }}
-            />
+            <div className={classes.buttonContainer}>
+                <Button
+                    text={"Criar prova"}
+                    onClick={props.changeCreateCompModal}
+                />
+            </div>
+            {props.createCompModal && (
+                <CreateCompetition
+                    createCompModal={props.createCompModal}
+                    changeCreateCompModal={props.changeCreateCompModal}
+                />
+            )}
         </div>
     );
 }
