+++ frontend/src/Components/Modals/CreateCompetition.module.css
@@ -1,0 +1,25 @@
+.createCompModalOverlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: rgba(0, 0, 0, 0.5);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    z-index: 89;
+}
+
+.createCompModal {
+    background: #fff;
+    padding: 20px;
+    border-radius: 8px;
+    text-align: center;
+    z-index: 90;
+}
+
+.createCompModal button {
+    margin-top: 10px;
+    z-index: 93;
+}
